fast/project-templates/secops-anonymization-pipeline/source/requirements.txt
# coding=utf-8
# Copyright 2025 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

google-cloud-pubsub
requests==2.27.1
jwt==1.3.1
google-auth
google-cloud-secret-manager
jinja2
google-cloud-storage
click==8.1.3
google-cloud-dlp
google-cloud-logging
